[#16600] Odyssey connection pooler CMake integration
Summary:
Odyssey is a PostgreSQL connection pooler. We already have Odyssey code as part of the yugabyte-db repository. This diff integrates Odyssey's CMake build scripts into the overall YugabyteDB build.

In this diff Odyssey build is disabled by default. It can be enabled on Linux by using the --with-odyssey option of yb_build.sh. Odyssey does not support macOS at the moment. Odyssey is also packaged by yb_release if it is built.

Other changes:
- Refactored build-support/yb_release.py into separate functions and renamed it to python/yb/yb_release_core_db.py.
- Introduced a mini-framework for "optional components" across yb_build.sh and yb_release (Python and Bash scripts). Currently there are two of these components: `odyssey` and `yugabyted_ui`. Whether or not to build them can be specified with `--with_<component>` and `--no_<component>` switches. They are disabled by default, except for yugabyted UI being enabled by default in yb_release (but not in yb_build.sh).
- Added shell scripts to codecheck.ini, but not enabling shellcheck in that file yet, because we need to install shellcheck on all build workers first.
- Added a more intelligent way to decide whether to automatically re-run CMake when certain switches given to yb_build.sh change compared to the corresponding values in CMakeCache.txt from the previous run of the build. The switches affected by this are currently --with-tests, --with-fuzz-targets, and --with-odyssey and their negative counterparts. The --with-yugabyted-ui is not relevant here because yugabyted UI build is invoked from Bash scripts and not from CMake.

Test Plan:
Jenkins

```
./yb_build.sh --clean --force --skip-java --with-odyssey --clang15 release
./yb_build.sh --clean --force --skip-java --with-odyssey --clang15 debug
./yb_build.sh --clean --force --skip-java --with-odyssey --clang15 fastdebug
./yb_build.sh --clean --force --skip-java --with-odyssey --clang15 release --lto=full
./yb_build.sh --clean --force --skip-java --with-odyssey --gcc11 release
./yb_build.sh --clean --force --skip-java --with-odyssey --gcc11 debug
./yb_build.sh --clean --force --skip-java --with-odyssey --gcc11 fastdebug
```

yb_release from scratch (without previously built code -- need to `rm -rf build` before every command):
```
./yb_release --with_odyssey --skip_yugabyted_ui_build --force
./yb_release --with_odyssey --skip_yugabyted_ui_build --force --build_args="--lto=full"
```

Check that the `odyssey` executable is present in the resulting archives:

```
tar tvf /path/to/yugabyte-a.b.c.d-githash-release-clang15-centos-x86_64.tar.gz | grep odyssey
```

Also test that yb_release is not broken without Odyssey:

```
./yb_release --skip_yugabyted_ui_build --force
./yb_release --skip_yugabyted_ui_build --force --build_args="--lto=full"
```

and check that the odyssey executable is not present in the resulting archives.

Reviewers: nkumar, janand, steve.varnau

Reviewed By: steve.varnau

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D23880

+++ python/yb/common_util.py
@@ -20,14 +20,14 @@ import logging
 import os
 import re
 import shutil
-import json
 import subprocess
 import shlex
 import io
 import platform
 import argparse
 import uuid
-import hashlib
+import pathlib
+import random
 
 import typing
 from typing import (
@@ -563,19 +563,28 @@ def assert_set_contains_all(a: Set[Any], b: Set[Any], message: str = '') -> None
             f"{set_to_str_one_element_per_line(d)}.")
 
 
-def create_temp_dir() -> str:
+def create_temp_dir(keep_tmp_dir: bool = False) -> str:
     """
-    Create a temporary directory and return its path.
-    The directory is automatically deleted at program exit.
+    Creates a temporary directory inside the "build" directory and returns its path.
+
+    :param keep_tmp_dir: If true, the temporary directory will not be deleted at exit.
     """
-    tmp_dir = os.path.join(YB_SRC_ROOT, "build", "yb_release_tmp_{}".format(str(uuid.uuid4())))
+
+    tmp_dir = os.path.join(
+        YB_SRC_ROOT,
+        "build",
+        "yb_tmp_{}_{}".format(
+            str(uuid.uuid4()),
+            ''.join([str(random.randint(0, 9)) for _ in range(16)])
+        )
+    )
     try:
-        os.mkdir(tmp_dir)
+        pathlib.Path(tmp_dir).mkdir(parents=True, exist_ok=True)
     except OSError as e:
         logging.error("Could not create directory at '{}'".format(tmp_dir))
         raise e
-
-    atexit.register(lambda: shutil.rmtree(tmp_dir))
+    if not keep_tmp_dir:
+        atexit.register(lambda: shutil.rmtree(tmp_dir))
     return tmp_dir
